Move page-level components to separate dir
