+++ app/scripts/pages/home.js
